requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-database-url==2.3.0
Django==5.2
django-allauth==65.7.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
